+++ src/lib/components/MapView/models.ts
@@ -1,1 +1,0 @@
-export type LngLatTuple = [lng: number, lat: number];