+++ src/lib/models/LngLatTuple.ts
@@ -1,0 +1,1 @@
+export type LatLngTuple = [lat: number, lng: number];